requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
jupyter
